fix time

+++ environments/mocaprasp/capture.py
@@ -31,7 +31,7 @@ picam2.set_controls(
         "AwbEnable": False,
         "Contrast": 32.0,
         # "ExposureValue": -2.0,
-        "ExposureTime": (1 / 10) * TIME_BUDGET * 1e6,
+        "ExposureTime": int((1 / 10) * TIME_BUDGET * 1e6),
     }
 )
